update1

+++ Entities/Concrete/Car.cs
@@ -12,7 +12,7 @@ namespace Entities.Concrete
         public int ColorId { get; set; }
         public int TransmissionId { get; set; }
         public string ModelYear { get; set; }
-        public double DailyPrice { get; set; }
+        public int DailyPrice { get; set; }
         public string Description { get; set; }
     }
 }
